made detector and classifier modules, combined detector with classifier, added FIXME for loading a trained classifier
