+++ deertracker/__init__.py
@@ -8,4 +8,7 @@ DEFAULT_PHOTO_STORE.mkdir(exist_ok=True)
 
 DEFAULT_DATABASE = DEFAULT_DATA_STORE / "deertracker.db"
 
-DEFAULT_MODEL_PATH = "models/md_v4.1.0.pb"
+DEFAULT_DETECTOR_PATH = "models/md_v4.1.0.pb"
+
+# FIXME:
+DEFAULT_CLASSIFIER_PATH = "models/#FIXME"
